+++ app/build.gradle
@@ -54,7 +54,7 @@ dependencies {
         exclude module: 'support-annotations'
     }
 
-    implementation 'com.github.TeamNewPipe:NewPipeExtractor:08f0247d7344'
+    implementation 'com.github.TeamNewPipe:NewPipeExtractor:94d1eb6f32'
 
     testImplementation 'junit:junit:4.12'
     testImplementation 'org.mockito:mockito-core:1.10.19'
